Update the Size type definition
Previously, for some misled and obscure reasons, the native size type
for the fbitset library was defined to be `int`.  This could limit the
size to be 512M bytes only.  And it may also implicitly introduce some
unnecessary casting when the native `size_t` is expected.  As a result,
in this commit, it is updated to the native `size_t`.

Although all test passes, all code using this library may have
compatibility issues.  This is the first, and probably only, backward
incompatible change that has ever happened to fbitset.

+++ include/fbitset.hpp
@@ -25,10 +25,9 @@ struct No_ext {
 
 /** The integral type used for bit/limb indices and sizes.
  *
- * Here we use int for performance and compatibility with the `digits` field of
- * numeric_limits.
+ * Here we just use the native size_t.
  */
-using Size = int;
+using Size = size_t;
 
 namespace internal {
     /** Utility for checking if the container is `No_ext`.
